Fix for large grids as contextual layers.

+++ src/main/java/au/org/ala/layers/intersect/Grid.java
@@ -548,8 +548,8 @@ public class Grid { //  implements Serializable
      */
     public void getClassInfo(Map<Float, float[]> info) {
 
-        long length = nrows * ncols;
-
+        long length = ((long) nrows) * ((long) ncols);
+        
         RandomAccessFile afile;
         File f2 = new File(filename + ".GRI");
